requirements.txt
python-dotenv
requests
billboard.py
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
pytube
moviepy
pillow==9.5.0
opencv-python
urllib3
tiktok-uploader
pydub